tools/grades/requirements.txt
playwright==1.48.0
python-dotenv
setuptools>=65.0.0
